# Enable sse2 for CSV parsing.